fix: Rename command folder to commands to match import statements
CRITICAL FIX: The folder was named 'command' but all imports throughout
the codebase use 'commands' (plural), causing import failures.

Changes:
- Rename command/ → commands/ to match expected imports
- Update documentation references to use correct path
- Update test imports to use commands module
- Verify imports work correctly with format_price test

This fixes a critical issue that would prevent the bot from running
due to ImportError on all command modules.
