+++ commands/auction_embed.py
