+++ commands/deal_confirmation.py
