+++ commands/sell_trade_embed.py
