+++ commands/trade.py
