+++ commands/trader_roles.py
